fix(ios): make BootSplash a single continuous screen from launch to hide
Previously the iOS launch sequence had two distinct splash surfaces with a
visible gap and size mismatch between them:

1. OS LaunchScreen renders BootSplash.storyboard out-of-process; logo sized
   to the imageView's intrinsic content size (108pt — the asset's natural
   size at 1×).
2. RCTBootSplash.initWithStoryboard was effectively never called because
   the AppDelegate cast `as? RCTRootView` returns nil on New Arch (the
   actual root is RCTSurfaceHostingProxyRootView, which inherits from
   RCTSurfaceHostingView, not RCTRootView). With no native overlay
   covering the React root view, the window's default white background
   showed through until React first-paint.
3. The JS SplashScreenHider then mounted with the logo hard-coded at 100pt
   (BootSplash.logoWidth fell back to its default because the native
   module's constantsToExport never exposed it). 108→100 = a visible
   7.4% shrink on top of the white flash.

This commit makes RCTBootSplash own the splash for the entire visible
lifetime, eliminating both the gap and the size jump:

- Pass `self.window?.rootViewController?.view` to RCTBootSplash as a plain
  UIView. RCTBootSplash.mm already re-casts to RCTSurfaceHostingProxyRootView
  internally; the old cast was the only thing skipping initialization on
  New Arch.
- Set window.backgroundColor to the storyboard's exact yellow so any frame
  where the storyboard subview is being laid out cannot peek white.
- Drop the `disableActivityIndicatorAutoHide:` and `setLoadingView:` calls
  in RCTBootSplash.mm. f618d0a2 attempted the same cast change and was
  reverted in 719c1c9c because of a stuck-forever bug pinned on those
  lifecycle hooks — on RN 0.81+ they register the loadingView with the
  React Surface lifecycle, which holds it until the surface itself reports
  ready. Plain addSubview: is enough; we don't need RN to manage the
  view's lifetime.
- Replace the JS SplashScreenHider with a direct BootSplash.hide() call
  from Kiroku.tsx. The native module performs the 250ms cross-dissolve
  in -hideAndClearPromiseQueue. The 15s safety timeout that lived in
  SplashScreenHider moves to Kiroku.tsx so we still log + force-hide if
  the JS gating condition (e.g. hasCheckedAutoLogin) ever deadlocks.
- Delete SplashScreenHider and the now-unused logoWidth/logoHeight/
  logoSizeRatio/navigationBarHeight fields from the BootSplash module.

Net result: the same storyboard UIView is on screen from app launch
through `BootSplash.hide()`, with the OS LaunchScreen handing off
seamlessly to the in-process subview. One fade at the end.
