+++ src/Kiroku.tsx
@@ -17,9 +17,9 @@ import * as User from './libs/actions/User';
 import Navigation from './libs/Navigation/Navigation';
 import NavigationRoot from './libs/Navigation/NavigationRoot';
 // import PushNotification from '@libs/Notification/PushNotification';
+import BootSplash from './libs/BootSplash';
 import Log from './libs/Log';
 import migrateOnyx from './libs/migrateOnyx';
-import SplashScreenHider from './components/SplashScreenHider';
 import * as ActiveClientManager from './libs/ActiveClientManager';
 import * as ErrorUtils from './libs/ErrorUtils';
 import * as UserUtils from './libs/UserUtils';
@@ -145,10 +145,44 @@ function Kiroku() {
     Navigation.setIsNavigationReady();
   }, []);
 
-  const onSplashHide = useCallback(() => {
-    setSplashScreenState(CONST.BOOT_SPLASH_STATE.HIDDEN);
-    // Performance.markEnd(CONST.TIMING.SIDEBAR_LOADED);
+  const hasHiddenSplash = useRef(false);
+
+  const hideSplash = useCallback(() => {
+    if (hasHiddenSplash.current) {
+      return;
+    }
+    hasHiddenSplash.current = true;
+    BootSplash.hide().finally(() => {
+      setSplashScreenState(CONST.BOOT_SPLASH_STATE.HIDDEN);
+    });
   }, [setSplashScreenState]);
+
+  // Drive the native RCTBootSplash hide once startup gating conditions are met.
+  // The native module performs the 250ms crossDissolve in
+  // ios/kiroku/RCTBootSplash.mm — we don't render a JS overlay anymore.
+  useEffect(() => {
+    if (!shouldHideSplash) {
+      return;
+    }
+    hideSplash();
+  }, [shouldHideSplash, hideSplash]);
+
+  // Safety net: if shouldHideSplash never becomes true (e.g. a startup gate
+  // deadlocks), force-hide so the user is never stuck on the splash overlay.
+  useEffect(() => {
+    const timeoutId = setTimeout(() => {
+      if (hasHiddenSplash.current) {
+        return;
+      }
+      Log.alert(
+        '[BootSplash] shouldHideSplash never became true, force-hiding splash',
+        {timeoutMs: 15 * 1000},
+        false,
+      );
+      hideSplash();
+    }, 15 * 1000);
+    return () => clearTimeout(timeoutId);
+  }, [hideSplash]);
 
   useLayoutEffect(() => {
     // Initialize this client as being an active client
@@ -310,12 +344,6 @@ function Kiroku() {
         lastVisitedPath={lastVisitedPath as Route}
         initialUrl={initialUrl}
       />
-      {splashScreenState !== CONST.BOOT_SPLASH_STATE.HIDDEN && (
-        <SplashScreenHider
-          shouldHideSplash={shouldHideSplash}
-          onHide={onSplashHide}
-        />
-      )}
     </>
   );
 }
