+++ src/libs/BootSplash/index.native.ts
@@ -11,8 +11,4 @@ function hide(): Promise<void> {
 export default {
   hide,
   getVisibilityStatus: BootSplash.getVisibilityStatus,
-  logoSizeRatio: BootSplash.logoSizeRatio || 1,
-  logoWidth: BootSplash.logoWidth || 100,
-  logoHeight: BootSplash.logoHeight || 100,
-  navigationBarHeight: BootSplash.navigationBarHeight || 0,
 };
