+++ src/libs/BootSplash/types.ts
@@ -1,10 +1,6 @@
 type VisibilityStatus = 'visible' | 'hidden';
 
 type BootSplashModule = {
-  logoSizeRatio: number;
-  logoWidth: number;
-  logoHeight: number;
-  navigationBarHeight: number;
   hide: () => Promise<void>;
   getVisibilityStatus: () => Promise<VisibilityStatus>;
 };
